@@ -8,7 +8,7 @@
     "test": "mocha"
   },
   "bin": {
-    "gnostic": "index.js"
+    "gnostic": "bin/gnostic.js"
   },
   "repository": {
     "type": "git",
@@ -25,24 +25,26 @@
     "debug": "^4.1.1",
     "esm": "^3.2.22",
     "lodash": "^4.17.11",
+    "ms": "^2.1.1",
     "rxjs": "^6.4.0",
     "traverse": "^0.6.6",
     "yargs": "^13.2.2"
   },
   "devDependencies": {
-    "eslint": "^5.15.3",
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.16.0",
     "eslint-config-prettier": "^4.1.0",
     "eslint-config-semistandard": "^13.0.0",
     "eslint-config-standard": "^12.0.0",
     "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-node": "^8.0.1",
     "eslint-plugin-prettier": "^3.0.1",
-    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-promise": "^4.1.1",
     "eslint-plugin-standard": "^4.0.0",
     "husky": "^1.3.1",
     "lint-staged": "^8.1.5",
     "markdownlint-cli": "^0.14.1",
-    "mocha": "^6.0.2",
+    "mocha": "^6.1.3",
     "prettier": "^1.16.4",
     "proxyquire": "^2.1.0",
     "sinon": "^7.3.1",
