@@ -1,7 +1,6 @@
-import {catchError, map} from 'rxjs/operators';
-
 import {bindNodeCallback} from 'rxjs';
 import fs from 'fs';
+import {map} from 'rxjs/operators';
 import {redact} from './redact';
 
 const readFile = bindNodeCallback(fs.readFile);
@@ -10,7 +9,5 @@ export const readReport = filepath =>
   readFile(filepath, 'utf8').pipe(
     map(JSON.parse),
     map(redact),
-    catchError(err => {
-      throw err;
-    })
+    map(Object.freeze)
   );
